refactor(lib): merge loader.ts into update.ts
